Refactoring authenticated link generation and moving it to a separate auth service
