+++ src/app.module.ts
@@ -12,6 +12,7 @@ import { APP_FILTER } from "@nestjs/core";
 import { GlobalHttpExceptionFilter } from "./common/filters/global-http-exception.filter";
 import { EventModule } from "./event/event.module";
 import { AttendedEventModule } from "./attended-event/attended-event.module";
+import { ResetPasswordModule } from './reset-password/reset-password.module';
 
 @Module({
   imports: [
@@ -24,7 +25,8 @@ import { AttendedEventModule } from "./attended-event/attended-event.module";
     UserModule,
     LoginModule,
     EventModule,
-    AttendedEventModule
+    AttendedEventModule,
+    ResetPasswordModule
   ],
   controllers: [AppController],
   providers: [
