+++ src/login/login.module.ts
@@ -9,6 +9,6 @@ import { AuthModule } from "../auth/auth.module";
 })
 export class LoginModule {
   configure(consumer: MiddlewareConsumer) {
-    consumer.apply(saveRedirectParam).forRoutes("login/google$");
+    consumer.apply(saveRedirectParam).forRoutes("login/google$", "login/facebook$");
   }
 }
