+++ views/login/index.ejs
@@ -43,6 +43,10 @@
     </div>
 
     <div class="text-center mt-3">
+      <a href="/auth/reset-password?r=<%= locals.redirectLink %>">Forgot your password?</a>
+    </div>
+
+    <div class="text-center mt-3">
       <a href="/auth/signup?r=<%= locals.redirectLink %>">Create an account</a>
     </div>
